Cleanup and organisation
Moved Utils around to be more descriptive
